# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EasyPrent Accounting",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/expenses": {
      "get": {
        "tags": [
          "Expenses"
        ],
        "summary": "List Expenses",
        "operationId": "expenses_list_expenses",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Expenses"
        ],
        "summary": "Create Expense",
        "operationId": "expenses_create_expense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/expenses/development": {
      "get": {
        "tags": [
          "Expenses"
        ],
        "summary": "Expense Development",
        "operationId": "expenses_get_expense_development",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9998,
              "minimum": 1900,
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDevelopmentResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/expenses/{expense_id}": {
      "put": {
        "tags": [
          "Expenses"
        ],
        "summary": "Update Expense",
        "operationId": "expenses_update_expense",
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Expense Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Expenses"
        ],
        "summary": "Delete Expense",
        "operationId": "expenses_delete_expense",
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Expense Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDeleteResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/expenses/{expense_id}/archive": {
      "post": {
        "tags": [
          "Expenses"
        ],
        "summary": "Archive Expense",
        "operationId": "expenses_archive_expense",
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Expense Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseLifecycleResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/expenses/{expense_id}/restore": {
      "post": {
        "tags": [
          "Expenses"
        ],
        "summary": "Restore Expense",
        "operationId": "expenses_restore_expense",
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Expense Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseLifecycleResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
...
    "/api/v1/settings/gnucash/accounts": {
      "get": {
        "tags": [
          "Settings and Backup"
        ],
        "summary": "Read Gnucash Accounts",
        "operationId": "settings_and_backup_read_gnucash_accounts",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GnuCashAccountResponse"
                  },
                  "type": "array",
                  "title": "Response Settings And Backup Read Gnucash Accounts"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "reason"
        ],
        "title": "ErrorDetail",
        "description": "A stable machine code and a readable explanation."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The common HTTP error envelope."
      },
      "ExpenseCategoryResponse": {
        "properties": {
          "expense_category": {
            "type": "string",
            "title": "Expense Category"
          },
          "beneficiary_name": {
            "type": "string",
            "title": "Beneficiary Name"
          },
          "expense_count": {
            "type": "integer",
            "title": "Expense Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expense_category",
          "beneficiary_name",
          "expense_count"
        ],
        "title": "ExpenseCategoryResponse"
      },
      "ExpenseDeleteResponse": {
        "properties": {
          "resource": {
            "type": "string",
            "const": "expenses",
            "title": "Resource"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resource",
          "id",
          "deleted"
        ],
        "title": "ExpenseDeleteResponse"
      },
      "ExpenseDevelopmentCategory": {
        "properties": {
          "expense_category": {
            "type": "string",
            "title": "Expense Category"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "has_uncalculated_expense": {
            "type": "boolean",
            "title": "Has Uncalculated Expense"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expense_category",
          "amount",
          "has_uncalculated_expense"
        ],
        "title": "ExpenseDevelopmentCategory"
      },
      "ExpenseDevelopmentMonth": {
        "properties": {
          "month": {
            "type": "integer",
            "title": "Month"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "has_uncalculated_expense": {
            "type": "boolean",
            "title": "Has Uncalculated Expense"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/ExpenseDevelopmentCategory"
            },
            "type": "array",
            "title": "Categories"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "month",
          "total_amount",
          "has_uncalculated_expense",
          "categories"
        ],
        "title": "ExpenseDevelopmentMonth"
      },
      "ExpenseDevelopmentResponse": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "has_uncalculated_expense": {
            "type": "boolean",
            "title": "Has Uncalculated Expense"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/ExpenseDevelopmentCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "months": {
            "items": {
              "$ref": "#/components/schemas/ExpenseDevelopmentMonth"
            },
            "type": "array",
            "title": "Months"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "year",
          "total_amount",
          "has_uncalculated_expense",
          "categories",
          "months"
        ],
        "title": "ExpenseDevelopmentResponse"
      },
      "ExpenseLifecycleResponse": {
        "properties": {
          "resource": {
            "type": "string",
            "const": "expenses",
            "title": "Resource"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resource",
          "id",
          "is_archived"
        ],
        "title": "ExpenseLifecycleResponse"
      },
      "ExpenseListResponse": {
        "properties": {
          "expenses": {
            "items": {
              "$ref": "#/components/schemas/ExpenseResponse"
            },
            "type": "array",
            "title": "Expenses"
          },
          "expense_categories": {
            "items": {
              "$ref": "#/components/schemas/ExpenseCategoryResponse"
            },
            "type": "array",
            "title": "Expense Categories"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expenses",
          "expense_categories"
        ],
        "title": "ExpenseListResponse"
      },
      "ExpenseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "object_type": {
            "type": "string",
            "enum": [
              "property",
              "building",
              "unit",
              "room"
            ],
            "title": "Object Type"
          },
          "object_id": {
            "type": "integer",
            "title": "Object Id"
          },
          "expense_category": {
            "type": "string",
            "title": "Expense Category"
          },
          "beneficiary_name": {
            "type": "string",
            "title": "Beneficiary Name"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "allocation_method": {
            "type": "string",
            "title": "Allocation Method"
          },
          "charge_type": {
            "type": "string",
            "enum": [
              "one_time",
              "monthly",
              "quarterly",
              "yearly",
              "consumption"
            ],
            "title": "Charge Type"
          },
          "recurrence": {
            "type": "string",
            "title": "Recurrence"
          },
          "interval": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval"
          },
          "meter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meter Id"
          },
          "meter_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meter Unit"
          },
          "consumption_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumption Unit"
          },
          "conversion_factor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversion Factor"
          },
          "consumption_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumption Value"
          },
          "effective_consumption_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Consumption Value"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "booking_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Date"
          },
          "period_start": {
            "type": "string",
            "title": "Period Start"
          },
          "period_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "is_open_ended": {
            "type": "boolean",
            "title": "Is Open Ended"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived",
            "default": false
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "object_type",
          "object_id",
          "expense_category",
          "beneficiary_name",
          "label",
          "amount",
          "allocation_method",
          "charge_type",
          "recurrence",
          "period_start",
          "is_open_ended"
        ],
        "title": "ExpenseResponse"
      },
      "ExpenseWrite": {
        "properties": {
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "object_type": {
            "type": "string",
            "enum": [
              "property",
              "building",
              "unit",
              "room"
            ],
            "title": "Object Type"
          },
          "object_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Object Id"
          },
          "expense_category": {
            "type": "string",
            "title": "Expense Category"
          },
          "beneficiary_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Name"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "allocation_method": {
            "type": "string",
            "title": "Allocation Method"
          },
          "charge_type": {
            "type": "string",
            "enum": [
              "one_time",
              "total",
              "monthly",
              "quarterly",
              "yearly",
              "consumption"
            ],
            "title": "Charge Type"
          },
          "booking_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Date"
          },
          "period_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "period_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "meter_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Meter Id"
          },
          "consumption_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumption Unit"
          },
          "consumption_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumption Value"
          },
          "conversion_factor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversion Factor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount",
          "object_type",
          "object_id",
          "expense_category",
          "allocation_method",
          "charge_type"
        ],
        "title": "ExpenseWrite"
      },
      "GnuCashAccountResponse": {
        "properties": {
          "guid": {
            "type": "string",
            "title": "Guid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "parent_guid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Guid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "guid",
          "name",
          "full_name",
          "parent_guid"
        ],
        "title": "GnuCashAccountResponse"
      },
...
    }
  }
}